Better integrating slicknav with more ugly hacks
We want the menu to expand fully and not close-on-click when the
screen size is larger than a certain amount. Have to make some
additional hacks to slicknav to make this happen. Should maybe
just do major work on slicknav at some point in the future. For
now, this will be good enough. We just need a menu, not good
javascript.