Learning the joys of having to reflow images for PDF output
So print output is SO finicky when it comes to images. For now,
I'm choosing to just manually move images in order to reflow print
output that comes from my machine using PrinceXML. Maybe there's a
better way to do it in the future.

My condolences to professional layout editors everywhere.