Adding CHANGELOG generation mechanism
Adding a shell script to generate a CHANGELOG file for each
document being built. It generates a markdown file from the git
commit history. This should be easier to link to and be more user
friendly for readers to see changes that were made between
releases.

This required some changes to the ant build file. Properties need
to now be passed around so that we can use a consistent timestamp.
That means removing antcall where necessary and adding inheritall
to the ant-contrib foreach tasks.

Requires git to be in the path. Currently no Windows batch file
yet.