Switching all books over to printhistory and pubdate elements
Now with pubdate support and the ability to display printhistory
in a book's titlepage, we should utilize that. Unfortunately, to
get a link to each book's respective CHANGELOG.md file, that means
we need a printhistory.xml per-book because the entities used for
the URL couldn't be overridden during XInclude as the document
needs to be parsed first (including entities).

Maybe there's a clever way around that, but I'm not sure of it.

Also restyling print output so that we get all the "bottom-matter"
together at the bottom of the printed titlepage.