Adding EPUB customization layer
Intentionally suppressing images from the EPUB output. The images
are unreliable when it comes to how it will fit on the display
screen, and it's possible the devices will be network-less or
show errors because they couldn't load the remote images. Maybe
we'll put them back in later, but for now it's safer to just remove
them from the output/