Adding customizations for epub titlepage output
Adding a custom titlepage template ALA
<http://www.sagehill.net/docbookxsl/HTMLTitlePage.html>
for EPUB output. The only real difference is that it prints the
`printhistory` element and puts a

> Last modified on

paragraph before the timestamp.

Added the ability to pass a timestamp into the EPUB stylesheets and
have that be used instead of the system time during transformation.
This also means that all timestamps have the same format, but I'm
fine with that.