Adding some custom CSS styling to EPUB output
Bringing over something similar to the HTML print CSS stylesheet
for EPUB styling. That way, if the user device doesn't override it,
we'll get something kind of the like the PDF but without columns
or pictures.

The styles are slightly different because the EPUB HTML comes from
the docbook-xsl-ns stylesheets rather than the xslt20 stylesheets.