Adding source SVG for first planned encounter of the scenario
Direction and plans for the encounter maps are still evolving.
For the time being, we think that fully labeled SVG images in the
books and in the repo is the way to go. Then, at release time,
we will produce an 'asset pack' from the source SVGs which may
include, but not be limited to:

* Unlabled SVG images
* Bitmap exports, perhaps high- and low-res