Preemptively adding a function to distinguish usable HTML images
As described in

https://github.com/rwdalpe/xslt20-stylesheets/commit/d55c4d2a05bdcb869630a23048f73317b95ff8c0

I have added the ability to customize which images are chosen for
use in a mediaobject element. Currently, the idea is that we'll be
able to choose between hires and lowres image options for the
various output types. Currently the code selects for lowres options
in HTML output, but that's undesirable long term.