Splitting out XML entities into separate files.
Eventually, I would probably like to take the `common/entities`
folder and move it to a separate repository so that any future
scenarios can all reference the same set of entities. At that
time, it will be necessary to have some sort of structure
implicitly defined on those entities.

This is my first attempt at this. The entities are broken down
by major geo-political region. Those are further broken down into
the elements that constitute that region.