Adding CLA to CONTRIBUTORS files, fixing calabash logging
Adding a contributor licensing agreemtn to the CONTRIBUTORS file
stating that submissions for inclusion in the project are agreed
to be released under the MIT License for code and the OGL for
textual content.

Calabash wants the log4j.xml configuration file to be in the
classpath before the jar files.