Enabling XML Catalog support and dynamic glossary
Utilizing my changes to jing to enable XML Catalog support.

This is necessary to enable a dynamic glossary as described in
http://www.sagehill.net/docbookxsl/GlossDatabase.html

I expect that I can use the catalog much more extensively, though
and that it will help with splitting out content that will be
shared between books into a separate repository in the future.

The glossary is obviously just for show at the moment and contains
nothing of value.

The EPUB build has not yet been updated to use the glossary.