Making catalog a little more robust with a build preparation step
Since various transformers and parsers might handle the catalog
differently, we'll make certain transformations a little more robust
by surrounding them in a group that specifies a base URL so that
we always get the right files.

To determine what the base URL should be, we add a build step to
set the base URL to the current path to the source files, and then
everything that references those files can be relative from that.