Changing out some glossary links for more appropriate elements
The player's guide currently has many glossary term references,
which are great, but there are more effective reference types that
could be used.

This commit swaps out some of the glossary references for `xrefs`
or `links` to internal sections of the player's guide.

This process is not yet complete and more commits on this matter
will follow.