Adding asciidoctor extension for person name formatting
Adding an extension that allows users to specify a format string
when processing a `person` macro.

The format string looks like

    '{f} {l} {0-n}'

And means the following:

All names are broken apart by spaces. For example, "Marshall Blackwood"
becomes [Marshall, Blackwood]

`{f}` will be replaced by the first component of the name string

`{l}` will be replaced by the last component of the name string

`{0-n}` will be replaced by the nth component of the name string

So `person::["Marshall Blackwood", "Emperor {l}"]` is replaced with

`<person><personname>Emperor Blackwood</personname></person>`

This is to address https://github.com/rwdalpe/two-graves/issues/8

And yes, this is obviously not a robus solution. `n` could be negative
and thus fail, and there's no real escaping method. If those ever
become issues, we'll address it then.