Adding the overland Norolba map to PG appendix
Included the overland Norolba map. This is significant for two reasons:

1. More images, yay!
2. It represents a first attempt at dynamic image rotation.

Full page images in print may need to be rotated for full effect. On the
web, though, they should not be rotated. This javascript/css/xslt will
display a particular image in print/non-js web, but when js is enabled
the image will be changed to the linked image. The linked image is
typically the non-rotated version.