Adding google analytics script to the HTML output customizations
Doing this so that I can gather some data about page visits.

For the privacy-concerned:

The google analytics script will not function when viewing HTML output
locally, because the page will be loaded in the file: scheme. The
analytics code uses //www.google-analytics.com when loading the script,
and of course file:///www.google-analytics.com doesn't exist.